Removed size param from addStorage
- The size was an easy way to create arrays, but since this value can be passed in the structName as an array amount this is not needed anymore
- Fixed all the examples by removing the extra parameters

+++ examples/params_test/structs.js
@@ -1,6 +1,6 @@
 export const structs = /*wgsl*/`
     struct Variable{
-        init: i32,
+        init: f32,
         color: vec4f,
     }
